Refactor IPC, Poll, Socket, and Library Structure for Scarlet Native Compatibility
- Removed legacy IPC implementation and replaced it with a compatibility facade for Scarlet Native IPC.
- Updated the library structure to integrate with scarlet_rt and scarlet_os, enhancing modularity and dependency management.
- Simplified Poll and Socket modules by replacing existing implementations with compatibility facades.
- Updated Cargo.lock files across sws-client and sws-protocol to reflect new dependencies and versions for scarlet-abi, scarlet-os, scarlet-rt, and scarlet-std.
- Ensured that all modules now utilize the new scarlet_os and scarlet_rt libraries for improved functionality and performance.
