+++ user/lib/scarlet-os/src/handle/capability/shared_memory.rs
@@ -15,7 +15,7 @@ pub enum SharedMemoryObjectError {
 
 /// Shared memory object capability
 use crate::handle::{Handle, RawHandle};
-use crate::syscall::{Syscall, syscall2};
+use scarlet_sys::{Syscall, syscall2};
 
 pub struct SharedMemoryObject<'a> {
     handle: &'a Handle,
