+++ user/lib/scarlet-os/src/handle/introspection.rs
