+++ user/lib/scarlet-os/src/hypervisor/types.rs
