+++ user/lib/scarlet-rt/src/arch/riscv64.rs
@@ -1,7 +1,11 @@
 use core::arch::{asm, naked_asm};
 
-use crate::{env, task::exit};
+use crate::{env, exit};
 
+/// Scarlet Native RISC-V entry trampoline.
+///
+/// This symbol lives in `.init` and branches to `_start`, matching the current
+/// Scarlet loader contract used by `scarlet_std`.
 #[unsafe(link_section = ".init")]
 #[unsafe(export_name = "_entry")]
 #[unsafe(naked)]
@@ -20,32 +24,45 @@ unsafe extern "Rust" {
     fn main() -> i32;
 }
 
+/// Scarlet Native RISC-V process entry point.
+///
+/// # Arguments
+///
+/// * `a0` - Argument count supplied by the Scarlet ABI loader.
+/// * `a1` - Argument vector pointer supplied by the Scarlet ABI loader.
+///
+/// # Safety
+///
+/// The loader must provide `a0` and `a1` according to the Scarlet Native
+/// process-start ABI.
 #[unsafe(link_section = ".init")]
 #[unsafe(export_name = "_start")]
-pub extern "C" fn _start(a0: usize, a1: usize) -> ! {
+pub unsafe extern "C" fn _start(a0: usize, a1: usize) -> ! {
     // Get argc and argv from RISC-V calling convention registers
     // a0 = argc, a1 = argv (set by kernel's ScarletAbi)
     let argc = a0;
     let argv = a1 as *const *const u8;
 
+    // Calculate envp from stack layout:
+    // Stack layout: argc | argv[] | envp[] | argv_strings | envp_strings
+    // envp starts right after argv[] (which has argc+1 entries including NULL)
+    let envp = if !argv.is_null() && argc > 0 {
+        // SAFETY: The Scarlet loader guarantees `argv` points to an argc-sized
+        // array followed by a null terminator and envp array.
+        unsafe { argv.add(argc + 1) }
+    } else {
+        core::ptr::null()
+    };
+
+    // SAFETY: `_start` is only entered by the Scarlet ABI loader, which sets up
+    // argc/argv/envp before transferring control.
     unsafe {
-        // Calculate envp from stack layout:
-        // Stack layout: argc | argv[] | envp[] | argv_strings | envp_strings
-        // envp starts right after argv[] (which has argc+1 entries including NULL)
-
-        // Handle NULL argv case - if argv is NULL, envp calculation is not safe
-        let envp = if !argv.is_null() && argc > 0 {
-            argv.add(argc + 1) as *const *const u8
-        } else {
-            core::ptr::null()
-        };
-
-        // Initialize environment before calling main
         env::init_env(argc, argv, envp);
-
-        let ret = main();
-        exit(ret as i32);
     }
+
+    // SAFETY: Scarlet `no_main` executables provide `extern "Rust" fn main`.
+    let ret = unsafe { main() };
+    exit(ret as i32);
 }
 
 /// Get the current thread's TLS (Thread Local Storage) pointer
@@ -84,5 +101,5 @@ pub fn arch_tls_pointer() -> usize {
 /// thread creation.
 pub fn arch_set_tls_pointer(ptr: usize) {
     // Use syscall to set TLS pointer (kernel needs to update ABI state)
-    crate::syscall::syscall1(crate::syscall::Syscall::SetTls, ptr);
+    scarlet_sys::syscall1(scarlet_sys::Syscall::SetTls, ptr);
 }
