+++ user/lib/std/Cargo.toml
@@ -14,10 +14,7 @@ alloc_debug = []
 [dependencies]
 scarlet-abi = { path = "../scarlet-abi" }
 scarlet-sys = { path = "../scarlet-sys" }
-talc = { version = "4.4", default-features = false, features = ["lock_api"] }
+scarlet-rt = { path = "../scarlet-rt", default-features = false, features = ["legacy-no-std"] }
+scarlet-os = { path = "../scarlet-os" }
 spin = { version = "0.10.0", default-features = false, features = ["lock_api", "mutex", "spin_mutex", "rwlock"] }
-lock_api = "0.4"
 hashbrown = "0.15"
-
-[build]
-target = "../../../targets/riscv64gc-unknown-scarlet-elf.json"
